requirements.txt
stoat.py
vk-audio-downloader